requirements.txt
anthropic>=0.25.0
sentence-transformers>=2.2.2
chromadb>=0.4.0
pymupdf>=1.23.0
pytesseract>=0.3.10
Pillow>=10.0.0
pdfplumber>=0.10.0
opencv-python>=4.8.0
python-docx>=0.8.11
numpy>=1.24.0
pandas>=2.0.0
streamlit>=1.28.0
python-dotenv>=1.0.0
tiktoken>=0.5.0